requirements.txt
streamlit
yt-dlp
whisper
elevenlabs
moviepy
python-dotenv
